rename behavior and change default one
