+++ docs/src/examples/components/Tooltip/BestPractices/TooltipBestPractices.tsx
@@ -4,6 +4,7 @@ import ComponentBestPractices from 'docs/src/components/ComponentBestPractices'
 
 const doList = [
   'Use `tooltipAsLabelBehavior` if adding tooltip to icon-only button or to another visual-only element without any text, title or label.',
+  'Use `tooltipAsDescriptionBehavior` if adding tooltip to a button or to another visual-only element with text, title or label.',
 ]
 
 const DropdownBestPractices: React.FunctionComponent<{}> = () => {
