+++ packages/accessibility/src/behaviors/index.ts
@@ -50,7 +50,7 @@ export { default as accordionBehavior } from './Accordion/accordionBehavior'
 export { default as accordionTitleBehavior } from './Accordion/accordionTitleBehavior'
 export { default as accordionContentBehavior } from './Accordion/accordionContentBehavior'
 export { default as checkboxBehavior } from './Checkbox/checkboxBehavior'
-export { default as tooltipBehavior } from './Tooltip/tooltipBehavior'
+export { default as tooltipAsDescriptionBehavior } from './Tooltip/tooltipAsDescriptionBehavior'
 export { default as tooltipAsLabelBehavior } from './Tooltip/tooltipAsLabelBehavior'
 export { default as sliderBehavior } from './Slider/sliderBehavior'
 export { default as menuButtonBehavior } from './MenuButton/menuButtonBehavior'
